Update WInterop
Update to .NET 10 (from .NET 6) and latest packages.

- TerraFX now uses char* instead of ushort* and new MethodCall calling convention
- xUnit now uses executable binaries
- Source generator now uses IIncrementalGenerator

+++ src/WInterop.Desktop/Registry/RegistryKeyHandle.cs
@@ -35,7 +35,7 @@ public class RegistryKeyHandle : HandleZeroOrMinusOneIsInvalid
     {
         // Need to cast to int first to avoid Overflow when
         // casting to IntPtr when running on 32 bit.
-        handle = (IntPtr)(int)predefined;
+        handle = (int)predefined;
         IsPerfKey = isPerfKey;
     }
